requirements.txt
Flask==2.1.3
flasgger==0.9.5
flask-cors==3.0.10
mysql-connector-python==8.0.33
azure-identity==1.13.0
azure-keyvault-secrets==4.5.1
Werkzeug==2.1.2
